requirements.txt
pytest==7.0.1
requests==2.28.2
requests_toolbelt==0.10.1
ConfigArgParse==1.5.3
wemake-python-styleguide==0.17.0
python-telegram-bot==13.15
python-dotenv==1.0.0